tree: distinguish between 'tls_key' and 'tls_configured_key'
The 'tls_key' sysfs attribute contains the negotiated TLS key
once the connection is encrypted; this key is ephemeral and might
change after each reset. So if a valid key is present here it
merely indicates that the connection is TLS encrypted.
The 'tls_configured_key' sysfs attribute contains the key the
user has specified via the '--tls_key' option during connect.
This key is fixed for the lifetime of the connection, and that
is the key which should be stored in the internal 'tls_key'
tree setting.

Signed-off-by: Hannes Reinecke <hare@suse.de>